+++ src/picker/js/constants.js
@@ -80,11 +80,8 @@ const MIME_TYPE_ALIAS = {
   ],
 };
 
-// set default error message timeout as 10 seconds,
-// and set error message timeout for computer uploads as false,
-// to keep all errors visible to users.
+// Set default error message timeout as 10 seconds.
 const ERROR_MSG_TIMEOUT = 10000;
-const UPLOAD_ERROR_MSG_TIMEOUT = false;
 
 export {
   TYPE_ALIAS,
@@ -92,7 +89,6 @@ export {
   VIEW,
   FLAVOR,
   ERROR_MSG_TIMEOUT,
-  UPLOAD_ERROR_MSG_TIMEOUT,
 };
 export default {
   TYPE_ALIAS,
@@ -100,5 +96,4 @@ export default {
   VIEW,
   FLAVOR,
   ERROR_MSG_TIMEOUT,
-  UPLOAD_ERROR_MSG_TIMEOUT,
 };