+++ src/picker/js/izitoast-helper.js
@@ -1,7 +1,5 @@
 /* eslint no-unused-vars: ["error", { "args": "none" }] */
 import iziToast from 'izitoast';
-import 'izitoast/dist/css/iziToast.css';
-import '../css/izitoast.less';
 import localization from './localization';
 import util from './util';
 import { ERROR_MSG_TIMEOUT } from './constants';
@@ -13,38 +11,18 @@ const supportsCopy =
 
 /**
  * Show an error toast.
+ * If there is detail message, a copy button is present and timeout is disabled.
  * @param {string} message
  * @param {object=} options
  * @param {string} options.detail - Detail message.
- * @param {number} options.timeout - Defaults to 1000ms. 0 to not close the
- * toast and clear old toasts as well.
  */
 function error(message, options = {}) {
   if (message === undefined || message === null || message === '') {
     return;
   }
-  const {
-    detail,
-    timeout = ERROR_MSG_TIMEOUT,
-    buttons = [],
-  } = options;
+  const { detail } = options;
 
-  const copyButton = [
-    `<button>${localization.formatAndWrapMessage('global/copy')}</button>`,
-    (instance, toast) => {
-      const msg = detail ? `${message} (detail: ${detail})` : message;
-      const eltemp = document.createElement('textarea');
-      eltemp.value = msg;
-      eltemp.readOnly = true;
-      eltemp.className = 'visual-invisible';
-      eltemp.setAttribute('aria-hidden', 'true');
-      document.body.appendChild(eltemp);
-      eltemp.select();
-      document.execCommand('copy');
-      document.body.removeChild(eltemp);
-    }, false,
-  ];
-
+  const buttons = [];
   let msg = message;
   if (detail) {
     if (util.isIE) {
@@ -54,22 +32,68 @@ function error(message, options = {}) {
       msg = `
       ${msg}
       <details>
-        <summary class="iziToast__details-toggle-button">show details</summary>
+        <summary class="iziToast__details-toggle-btn">show details</summary>
         ${detail}
       </details>`;
     }
+    if (supportsCopy) {
+      buttons.push([
+        `<button class="iziToast__copy-btn">
+          ${localization.formatAndWrapMessage('global/copy')}
+        </button>`,
+        (instance, toast) => {
+          const eltemp = document.createElement('textarea');
+          eltemp.value = `${message} (detail: ${detail})`;
+          eltemp.readOnly = true;
+          eltemp.className = 'visual-invisible';
+          eltemp.setAttribute('aria-hidden', 'true');
+          document.body.appendChild(eltemp);
+          eltemp.select();
+          document.execCommand('copy');
+          document.body.removeChild(eltemp);
+        }, false,
+      ]);
+    }
   }
-  if (timeout === 0) {
-    iziToast.destroy();
-  }
-  iziToast.error({
+
+  iziToast.show({
+    buttons,
+    progressBar: !detail,
+    timeout: detail ? false : ERROR_MSG_TIMEOUT,
+    layout: detail ? 2 : 1, // 1: small layout, 2: medium layout
     pauseOnHover: true,
     drag: false, // to allow user to select text
-    timeout: timeout || false,
     position: 'bottomCenter',
     title: 'Error',
     message: msg,
-    buttons: supportsCopy ? [copyButton, ...buttons] : [...buttons],
+    theme: 'custom-error',
+    icon: 'material-icons',
+    iconText: 'highlight_off',
+  });
+}
+
+function success(message) {
+  iziToast.show({
+    close: false,
+    drag: false, // to allow user to select text
+    timeout: false,
+    position: 'center',
+    title: 'Success',
+    message,
+    progressBar: false,
+    layout: 1,
+    overlay: true,
+    icon: 'material-icons',
+    iconText: 'check_circle',
+    theme: 'custom-success',
+    displayMode: 'replace',
+    buttons: [[
+      `<button class="iziToast__ok-btn">
+        ${localization.formatAndWrapMessage('global/ok')}
+      </button>`,
+      (instance, toast) => { instance.hide({}, toast); },
+      false,
+    ]],
   });
 }
 
@@ -80,4 +104,4 @@ function destroy() {
   iziToast.destroy();
 }
 
-export default { error, destroy };
+export default { error, destroy, success };